+++ src/common/symplug.c
@@ -37,6 +37,9 @@ static void symplug_join(void *handle, void *calls[], const char *names[], uint 
 
 	for (i = 0; i < n; ++i) {
 		calls[i] = dlsym(handle, names[i]);
+		if (calls[i]!=NULL) {
+			// fprintf(stderr,"symbol %s found\n",names[i]);
+		}
 	}
 }