+++ src/common/time.c
@@ -65,6 +65,7 @@ ullong timestamp_diff(timestamp *ts2, timestamp *ts1, ullong time_unit)
 	ullong stamp;
 
 	if (ts2->tv_nsec < ts1->tv_nsec) {
+		ts2->tv_sec   = ts2->tv_sec - 1;
 		ts2->tv_nsec += 1000000000;
 	}
 
@@ -73,4 +74,13 @@ ullong timestamp_diff(timestamp *ts2, timestamp *ts1, ullong time_unit)
 	stamp /= time_unit;
 
 	return stamp;
+}
+
+int main(int argc, char *argv[])
+{
+	timestamp_t
+	ullong diff;
+
+	timestamp_getfast(&t1);
+	fus = timestamp_diff(&ts2, &ts1, TIME_USECS);
 }