+++ src/common/types/configuration/cluster_conf.c
@@ -234,6 +234,7 @@ my_node_conf_t *get_my_node_conf(cluster_conf_t *my_conf,char *nodename)
 	int i=0, j=0;
 	my_node_conf_t *n=calloc(1, sizeof(my_node_conf_t));
     n->num_policies = my_conf->num_policies;
+		n->policies=malloc(sizeof(policy_conf_t)*n->num_policies);
     int num_spec_nodes = 0;
     int range_id = -1;
     while(i<my_conf->num_nodes)
@@ -273,11 +274,16 @@ my_node_conf_t *get_my_node_conf(cluster_conf_t *my_conf,char *nodename)
 
 
     //pending checks for policies
+    fprintf(stderr,"Initialzing %u policies for node %s,%u special policies \n",my_conf->num_policies,nodename,
+		num_spec_nodes);
+		memcpy(n->policies,my_conf->power_policies,sizeof(policy_conf_t)*my_conf->num_policies);
+		#if 0
     for (i = 0; i < my_conf->num_policies; i++)
     {
         char found = 0;
-        for (j = 0; j < num_spec_nodes; j++)
+        for (j = 0; j < num_spec_nodes; j++){
             if (my_conf->power_policies[i].policy == n->policies[j].policy) found = 1;
+				}
 
         if (!found)
         {
@@ -285,6 +291,7 @@ my_node_conf_t *get_my_node_conf(cluster_conf_t *my_conf,char *nodename)
             num_spec_nodes++;
         }
     }
+		#endif
 	n->max_pstate=my_conf->eard.max_pstate;
 
 	return n;
@@ -590,7 +597,7 @@ void set_default_island_conf(node_island_t *isl_conf, uint id)
 void set_default_conf_install(conf_install_t *inst)
 {
 	sprintf(inst->obj_ener, "default");
-	sprintf(inst->obj_poli, "default");
+	sprintf(inst->obj_power_model, "default");
 }
 
 /*