+++ src/library/models/basic_model.c
@@ -30,11 +30,12 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <common/config.h>
-#include <common/output/verbose.h>
 #include <common/states.h>
 #include <common/types/signature.h>
 #include <daemon/shared_configuration.h>
 #include <control/frequency.h>
+
+#include <library/models/models_api.h>
 
 
 static coefficient_t **coefficients;
@@ -51,7 +52,7 @@ static int valid_range(ulong from,ulong to)
 }
 
 /* This function loads any information needed by the energy model */
-state_t ear_model_init(char *etc,char *tmp,uint pstates)
+state_t model_init(char *etc,char *tmp,uint pstates)
 {
   char coeff_file[128];
   char coeff_default_file[128];
@@ -63,13 +64,13 @@ state_t ear_model_init(char *etc,char *tmp,uint pstates)
 
   coefficients = (coefficient_t **) malloc(sizeof(coefficient_t *) * num_pstates);
   if (coefficients == NULL) {
-    error("Error allocating memory for p_states coefficients\n");
+		return EAR_ERROR;
   }
   for (i = 0; i < num_pstates; i++)
   {
     coefficients[i] = (coefficient_t *) malloc(sizeof(coefficient_t) * num_pstates);
     if (coefficients[i] == NULL) {
-      error("Error allocating memory for p_states coefficients fn %d\n",i);
+			return EAR_ERROR;
     }
 
     for (ref = 0; ref < num_pstates; ref++)
@@ -94,15 +95,8 @@ state_t ear_model_init(char *etc,char *tmp,uint pstates)
       i=frequency_freq_to_pstate(coefficients_sm[ccoeff].pstate);
       if (frequency_is_valid_pstate(ref) && frequency_is_valid_pstate(i)){
 				memcpy(&coefficients[ref][i],&coefficients_sm[ccoeff],sizeof(coefficient_t));
-      }else{
-        if (coefficients_sm[ccoeff].available){
-          verbose(0,"Error: invalid coefficients for ref %lu or proj %lu\n",coefficients_sm[ccoeff].pstate_ref,
-          coefficients_sm[ccoeff].pstate);
-        }
       }
     }
-  }else{
-    verbose(1,"NO coefficients found");
   }
 	basic_model_init=1;	
 	return EAR_SUCCESS;
@@ -115,7 +109,7 @@ double default_project_cpi(signature_t *sign, coefficient_t *coeff)
 		   (coeff->F);
 }
 
-state_t ear_model_project_time(signature_t *sign,ulong from,ulong to,double *ptime)
+state_t model_project_time(signature_t *sign,ulong from,ulong to,double *ptime)
 {
 	state_t st=EAR_SUCCESS;
 	coefficient_t *coeff;
@@ -136,7 +130,7 @@ state_t ear_model_project_time(signature_t *sign,ulong from,ulong to,double *pti
 	return st;
 }
 
-state_t ear_model_project_power(signature_t *sign, ulong from,ulong to,double *ppower)
+state_t model_project_power(signature_t *sign, ulong from,ulong to,double *ppower)
 {
 	state_t st=EAR_SUCCESS;
 	coefficient_t *coeff;
@@ -154,7 +148,7 @@ state_t ear_model_project_power(signature_t *sign, ulong from,ulong to,double *p
 	return st;
 }
 
-state_t ear_model_projection_available(ulong from,ulong to)
+state_t model_projection_available(ulong from,ulong to)
 {
 	if (coefficients[from][to].available) return EAR_SUCCESS;
 	else EAR_ERROR;