+++ src/slurm_plugin/slurm_plugin_rcom_eargmd.c
@@ -37,13 +37,21 @@ int plug_rcom_eargmd_job_start(spank_t sp, plug_serialization_t *sd)
 	plug_verbose(sp, 2, "function plug_rcom_eargmd_job_start");
 
 	// Disabled
-	return ESPANK_SUCCESS;
+	//return ESPANK_SUCCESS;
 
 	// Pack deserialization
-	getenv_agnostic(sp, "EARGMD_CONNECTED", buffer, SZ_PATH);
-	sd->pack.eargmd.connected = atoi(buffer);
+	sprintf(buffer, "\0");
+	getenv_agnostic(sp, Var.gm_secure.loc, buffer, SZ_PATH);
+	sd->pack.eargmd.secured = atoi(buffer);
 
-	if (sd->pack.eargmd.enabled && !sd->pack.eargmd.connected) {
+	if (!sd->pack.eargmd.enabled || sd->pack.eargmd.secured) {
+		plug_verbose(sp, 3, "connection with EARGMD not enabled or secured");
+		return ESPANK_SUCCESS;
+	}
+
+	//
+	if (sd->job.n_nodes == 0) {
+		plug_error(sp, "while getting the node number '%u'", sd->job.n_nodes);
 		return ESPANK_ERROR;
 	}
 
@@ -64,8 +72,8 @@ int plug_rcom_eargmd_job_start(spank_t sp, plug_serialization_t *sd)
 	// Informing that this report has to be finished
 	sd->pack.eargmd.connected = 1;
 
-	// Enabling protection
-	setenv_agnostic(sp, "EARGMD_CONNECTED", "1", 1);
+	// Enabling protection (hypercontained)
+	setenv_agnostic(sp, Var.gm_secure.loc, "1", 1);
 
 	return ESPANK_SUCCESS;
 }
@@ -75,7 +83,14 @@ int plug_rcom_eargmd_job_finish(spank_t sp, plug_serialization_t *sd)
 	plug_verbose(sp, 2, "function plug_rcom_eargmd_job_finish");
 
 	// Disabled
-	return ESPANK_SUCCESS;
+	//return ESPANK_SUCCESS;
+
+	if (!sd->pack.eargmd.connected) {
+		return ESPANK_SUCCESS;
+	}
+
+	plug_verbose(sp, 2, "trying to disconnect EARGMD with host '%s', port '%u', and nnodes '%u'",
+                sd->pack.eargmd.host, sd->pack.eargmd.port, sd->job.n_nodes);
 
 	if (eargm_connect(sd->pack.eargmd.host, sd->pack.eargmd.port) < 0) {
 		plug_error(sp, "while connecting with EAR global manager daemon");
@@ -87,8 +102,8 @@ int plug_rcom_eargmd_job_finish(spank_t sp, plug_serialization_t *sd)
 	}
 	eargm_disconnect();
 
-	// Disabling protection
-	setenv_agnostic(sp, "EARGMD_CONNECTED", "0", 1);
+	// Disabling protection (hypercontained)
+	setenv_agnostic(sp, Var.gm_secure.loc, "0", 1);
 
 	return ESPANK_SUCCESS;
 }